Creact CopyToClipboard component
